Update seccomp profile
The profile is now based on the Docker default profile[0] as of this
commit[1].

This is intended to resolve issues where newer container images crashed
using the previous profile due to issues like `clone3` not being
supported properly (which had been fixed upstream[2]).

Most likely, we will need to continue manually updating this profile in
the future to avoid similar incompatibility issues. Initially, I tried
seeing whether we could retrieve the default profile directly from the
Docker package and apply our desired changes on top, but this is not
possible while retaining OS cross-compatibility, since the file that
exports the default profile is only compiled on Linux[3].

So, we'll most likely have to keep regenerating this file by hand from
time to time. I've started a practice of adding a "cmgr: " comment
to our customizations (inspired by this[4] upstream commit, which was
later reverted[5], possibly by accident?). If we continue this policy,
then the process of updating this file can be as simple as:

1. Retrieving the latest upstream default seccomp profile[0]
2. Merging in any statements with "cmgr: " comments from the existing file
3. Replacing the file with the new version

Currently, the only customizations are to allow any combination of the
UNAME26, ADDR_NO_RANDOMIZE, and PER_LINUX32 flags to the personality()
syscall.

[0]: https://github.com/moby/moby/blob/master/profiles/seccomp/default.json
[1]: https://github.com/moby/moby/commit/891241e7e74d4aae6de5f6125574eb994f25e169
[2]: https://github.com/moby/moby/commit/9f6b562dd12ef7b1f9e2f8e6f2ab6477790a6594
[3]: https://github.com/moby/moby/blob/master/profiles/seccomp/seccomp_linux.go
[4]: https://github.com/ArmyCyberInstitute/cmgr/commit/9f7c6868e9b177b34003fd3c04b045f6337d1b09
[5]: https://github.com/ArmyCyberInstitute/cmgr/commit/bbae80d6d2f6f027b4e7b6f74c469569d8a9bf0d