Remove `black` formatting check for Python code
Currently failing in CI because the styles enforced by the `black` apt
package differ between Ubuntu 20.04 and Ubuntu 22.04 (ubuntu-latest).

This will likely continue to be an issue going forward if we want to
support historical Ubuntu LTS versions and I'm not sure it's worth the
hassle to fix (we could install from PyPI instead).

Possibly in the future we could use `ruff` instead, which has
essentially superceded `black` as the recommended Python formatter.