Remove support for most templates
`remote-make` and `static-make` are still supported, only because we
have a large number of challenges still using these templates in
production. We have found that templated challenges are difficult to
debug due to the indirection from the underlying Dockerfile. In
additional, several of the original templates no longer work due to
dependency drift. In order to reduce the maintenance burden, we are
removing all templates other the `remote-make` and `static-make`, and
make remove those as well in the future after converting existing
challenges to `custom`.